+++ src/mc/mc_location.rs
@@ -1,9 +1,9 @@
 /// Structure used to divide and manage physical space of the problem.
 #[derive(Debug, PartialEq)]
 pub struct MCLocation {
-    pub domain: u32, // u32? usize? usize would be good but we need a special default value; usize::MAX?
-    pub cell: u32,
-    pub facet: u32,
+    pub domain: usize, // u32? usize? usize would be good but we need a special default value; usize::MAX?
+    pub cell: usize,
+    pub facet: usize,
 }
 
 impl Default for MCLocation {
