+++ src/mc/mc_nearest_facet.rs
@@ -5,7 +5,7 @@ use num::Float;
 #[derive(Debug, Clone)]
 pub struct MCNearestFacet<T: Float> {
     /// Facet the particle is the closest to
-    pub facet: u32,
+    pub facet: usize,
     /// Distance between facet and particle
     pub distance_to_facet: T,
     /// Dot product between facet and speed/position vector?
